Adapt v2 pipeline for single A800-80G (bf16, no-offload, concurrent annotate)
- training: bf16=True/fp16=False for SFT+DPO; bigger micro-batch with
  smaller grad-accum (effective batch unchanged: SFT 32, DPO 16)
- configs: add ds_zero2.json (ZeRO-2, no CPU offload, bf16) — 3B full
  fine-tune fits on 80G; offloading to the 18-core CPU was the V100 path
- sampling: dtype float16 -> bfloat16 to match StarCoder2 weights/training
- annotate: batch_annotate now thread-pooled (bounded in-flight, single
  writer, exception-safe final=-1, enlarged conn pool); add --concurrency
  (default 50, GLM-4-Air allows ~100). Reuses judge_client backoff retry
- README: A800 single-card minimal-loop deploy notes (ds_zero2, --concurrency,
  skip --do_timing so the 18-core CPU is not the bottleneck)

+++ v2/scripts/02_sample_pilot.py
@@ -133,7 +133,8 @@ def main():
     log.info(f"selected {len(chosen)} problems for pilot")
 
     # 2) Sketch 采样
-    backend = build_backend(args.model_path, dtype="float16", prefer=args.prefer_backend)
+    # A800: bfloat16,与 StarCoder2 原生权重 + bf16 训练一致
+    backend = build_backend(args.model_path, dtype="bfloat16", prefer=args.prefer_backend)
 
     # 拿个 tokenizer 做长度过滤(直接用 backend 内部的)
     tokenizer = getattr(backend, "tokenizer", None)
